+++ trunk/src/dmcigd/core/objects/interfaces/Item.java
@@ -2,8 +2,8 @@ package dmcigd.core.objects.interfaces;
 
 public interface Item extends SolidObject {
 	
-	public void setX(int x);
-	public void setY(int y);
+	public void setX(float x);
+	public void setY(float y);
 	public void setVX(float vx);
 	public void setVY(float vy);
